==> src/01 Preprocessing.ipynb <==
---
title: Preprocessing
description: Preprocessing the data for future use
---

from pathlib import Path

import pandas as pd
import polars as pl
from IPython.display import HTML, display

### Preprocessing match results

# | label: match_results
# Column names : 'league', 'country', 'season', 'date', 'home', 'home_goals', 'away', 'away_goals'
match_results = (
    pl.read_csv(
        Path("./../data/extracted_match_results.csv"),
        infer_schema=True,
        infer_schema_length=10000,
        try_parse_dates=True,
        null_values=["NA"],
    )
    .cast(
        {
            "home_goals": pl.Int64,
            "away_goals": pl.Int64,
        }
    )
    .with_columns(
        pl.col("country").replace(
            {
                "ENG": "England",
                "ITA": "Italy",
                "FRA": "France",
                "GER": "Germany",
                "ESP": "Spain",
            }
        )
    )
    .drop_nulls(subset=["home_goals", "away_goals"])
)

match_results.head()

match_results.describe()

def determine_result(goals: pl.Expr, opponent_goals: pl.Expr) -> pl.Expr:
    """Determine match result from goals scored perspective."""
    return (
        pl.when(goals > opponent_goals)
        .then(pl.lit("win"))
        .when(goals < opponent_goals)
        .then(pl.lit("loss"))
        .otherwise(pl.lit("draw"))
    )


# Add match results and reshape data
match_results = match_results.with_columns(
    determine_result(pl.col("home_goals"), pl.col("away_goals")).alias("home_result"),
    determine_result(pl.col("away_goals"), pl.col("home_goals")).alias("away_result"),
)

# Combine home and away results into single format
match_results = pl.concat(
    [
        match_results.select(
            "league",
            "country",
            "date",
            pl.col("home").alias("team"),
            pl.col("home_goals").alias("goals"),
            pl.col("home_result").alias("result"),
            pl.lit(True).alias("is_home"),
        ),
        match_results.select(
            "league",
            "country",
            "date",
            pl.col("away").alias("team"),
            pl.col("away_goals").alias("goals"),
            pl.col("away_result").alias("result"),
            pl.lit(False).alias("is_home"),
        ),
    ],
    how="vertical",
)

match_results

### Preprocessing head coach

# | label: head_coach
# Column names : 'Team', 'League', 'Country', 'HeadCoach', 'Appointed', 'EndDate', 'Tenure', 'Matches', 'Wins', 'Draws', 'Losses'
head_coach = pl.read_csv(
    Path("./../data/extracted_head_coach.csv"),
    infer_schema=True,
    infer_schema_length=10000,
    try_parse_dates=True,
)
head_coach.head()

head_coach.describe()

We need to filter head coach that were not active between 2015 to 2023.

latest_match = match_results.get_column("Date").max()
earliest_match = match_results.get_column("Date").min()

# Remove head coach that were appointed after 2023 season
head_coach = head_coach.filter(pl.col("Appointed") <= latest_match)
# Keep head coach that were dimissed after 2015 or that are still active
head_coach = head_coach.filter(
    (pl.col("EndDate") >= earliest_match) | (pl.col("EndDate").is_null())
)

:::{caution}
Head coach appointments records, extracted from TransferMarkt, contains data related to head coach in that specific club : tenure, number of matches played, number of matches won, etc. Those datapoint goes beyond our cut-off date of 2023 end season (2024-01-14).

One important thing is that those feature beyond cut-off date still relate to a head coach appointment we have in our records. This guarantees us that metrics such *number of club head coach managed* are properly reflected and still relate to head coach performance.

However, this create an asymetry in our data, as certain data point are limited by a time-frame and others not.
Moreover, we must be careful in how we compare these datapoint to others dataset such as match results as it could easily bias our statistical study.

The only way we combine this dataset to match result is by extracting head coach tenure on day of a match. This does not bias our statistical study as it is a feature that is properly reflected by our cut-off date.
:::

:::{note}
I have considered computing Head Coach performance metric from match results but we would lose information on prior records as well as creating imbalance data for plot such as linear regression of head coach performance over head coach tenure : a long standing coach which would not been dismissed soon after our start date would have a lower number of matches, thus a performance metric with higher variance that would bias linear regression due to long tenure.
:::

# Ensuring there is only 1 head coach at a time in any given team.
head_coach_temp = head_coach.copy()

# Sort data by 'Team' and 'Appointed'
head_coach_temp = head_coach_temp.sort_values(["Team", "Appointed"])
# Fillna with end date of 2022-2023 season
head_coach_temp = head_coach_temp.with_columns(pl.col("EndDate").fillna("2024-01-14"))
# Check if the next appointment is overlapping with the current one
head_coach_temp["OverlapDuration"] = (
    head_coach_temp.groupby("Team")["Appointed"].shift(-1) - head_coach_temp["EndDate"]
)
head_coach_temp = head_coach_temp.with_columns(
    Overlap=head_coach_temp["OverlapDuration"].dt.days < 0
)
# Show team with overlapping appointments
overlapping = head_coach_temp.filter(pl.col("Overlap"))

Nous avons trouvé {eval}`overlapping.shape[0]` enregistrements de mandats d'entraîneurs sportifs qui avaient lieu alors qu'un autre chef-entraîneur assurait l'entraînement de l'équipe. Ces enregistrements concernent les équipes suivante : {eval}`', '.join(overlapping['Team'].unique())`. Ces enregistrements sont exclus du jeu de données.

head_coach_temp.sort_values("OverlapDuration").head()

# Show overlapping records for teams with overlaps, using polars
head_coach_temp.filter(pl.col("Team").is_in(overlapping["Team"])).sort(
    ["Team", "Appointed"]
).select(["Team", "Appointed", "EndDate", "Overlap", "OverlapDuration"]).head(10)

# | label: hc_inconsistency

head_coach_temp.filter(pl.col("Team") == "Stade Reims").sort("Appointed").select(
    ["Team", "Appointed", "EndDate", "Overlap"]
).slice(2, 4)

# Check records that are overlapping
head_coach[
    ~head_coach.index.isin(
        head_coach_temp[head_coach_temp["OverlapDuration"].dt.days <= -20].index
    )
]
# Drop overlapping records with a duration of more than 20 days
head_coach = head_coach[
    head_coach.index.isin(
        head_coach_temp[head_coach_temp["OverlapDuration"].dt.days > -20].index
    )
]

Next we would like to add to each head coach record the number of appointment he is completing.

# Using cronological information about appointment
# Add a column to head coach records that tells us about how many appointment head coach has done

head_coach = head_coach.sort_values(["HeadCoach", "Appointed"])
head_coach["AppointmentNumber"] = head_coach.groupby("HeadCoach").cumcount() + 1
head_coach[
    head_coach["HeadCoach"].isin(
        head_coach[head_coach["AppointmentNumber"] > 3]["HeadCoach"]
    )
].sort_values("HeadCoach").head()

# Check if total_matches = wins + draws + losses
head_coach[
    head_coach["Matches"]
    != head_coach["Wins"] + head_coach["Draws"] + head_coach["Losses"]
].shape[0]

# Display Head Coach with lowest number of match
display(head_coach.sort_values("Matches").head(20))
# Remove head coach records with less than 5 matches
# We consider them to be not relevant as it a short stay of a head coach indicate either a temporary replacement or a very bad performance
head_coach = head_coach[head_coach["Matches"] > 5]

#### Joining head coach with match results

# Compute number of team that are in head_coach but not in match_results
coach_teams = set(head_coach["Team"])
match_teams = set(match_results["Team"])

coach_team_not_in_match = coach_teams - match_teams
match_team_not_in_coach = match_teams - coach_teams

len(coach_team_not_in_match), len(match_team_not_in_coach)

Les résultats de matchs contiennent {eval}`len(match_teams)` équipes et les mandats de coachs contiennent {eval}`len(coach_teams)` équipes. Cependant, le nom de certaines équipes est différent entre les deux jeux de données. Par exemple, « Liverpool » dans les résultats de match devient « Liverpool FC » dans les mandats des entraîneurs.

Il existe {eval}`len(coach_team_not_in_match)` équipes présentes dans les mandats de coachs qui n'ont pas de correspondance dans les résultats de match.

# Group teams by country
coach_teams_by_country = head_coach.groupby("Country")["Team"].unique()
match_teams_by_country = match_results.groupby("Country")["Team"].unique()

coach_teams_by_country, match_teams_by_country

from thefuzz import process

team_name_mapping = {}

# For each country
for country in coach_teams_by_country.index:
    # Get teams for this country
    coach_teams = coach_teams_by_country[country]
    match_teams = match_teams_by_country.get(country, [])

    # For each team in coach_teams
    for coach_team in coach_teams:
        # Find the best match in match_teams
        matching_scores = process.extract(coach_team, match_teams, limit=1)

        if len(matching_scores) != 0 and matching_scores[0][1] >= 60:
            team_name_mapping[coach_team] = matching_scores[0][0]
        else:
            team_name_mapping[coach_team] = None
            print(f"No match found for {coach_team} among {match_teams} in {country}")

# | label: team_match_table

name_match = pd.DataFrame(
    team_name_mapping.items(),
    columns=["Team in head coach records", "Team in match results"],
)
display(HTML(name_match.head().to_html(index=False)))

# Map head_coach['team'] with name_match
head_coach["Team"] = head_coach["Team"].map(team_name_mapping)
head_coach.head()

Nous pouvons maintenant ajouter le nombres de jours au poste d’entraîneur avec les résultats de match.

# Check match with NaN goals
display(match_results[match_results["Goals"].isna()])
# Remove match with NaN goals
match_results = match_results[~match_results["Goals"].isna()]

# | label: final_match_results

# Merge head_coach with match_results
match = match_results.merge(
    head_coach[["Team", "HeadCoach", "Appointed", "EndDate"]], on=["Team"], how="left"
)
# Put aside team that don't have a head coach
no_headcoach = match[match["HeadCoach"].isna()]
match = match[~match["HeadCoach"].isna()]
# Filter match_results_bis to keep only head coach that were appointed before the match and with no end date or end date after the match
match = match[
    (match["Date"] >= match["Appointed"])
    & ((match["Date"] <= match["EndDate"]) | match["EndDate"].isna())
]
# Join back the team that don't have a head coach
match = pd.concat([match, no_headcoach], ignore_index=True)
# Compute daysInPost
match["DaysInPost"] = (match["Date"] - match["Appointed"]).dt.days
match = match.drop(columns=["Appointed", "EndDate"])
match.head()

## Missing head coach data 

Il existe des matchs sur les lesquels nous n'avons pas d'information sur le coach en poste.

# Values count of coach record / no coach record per team
team_hc_count = match.groupby("Team")["HeadCoach"].count()
# Display list of team with no head coach record
team_no_hc_record = team_hc_count[team_hc_count == 0].index.tolist()
print(len(team_no_hc_record))
print(team_no_hc_record)

### Saving preprocessed data

# Save match_results
match.to_csv(Path("./../data/match_results.csv"), index=False)
head_coach.to_csv(Path("./../data/head_coach.csv"), index=False)
